A little more test refactoring
Moved machine-specific subroutines into machines.py, moved pause() into
general.py
