+++ tests/python_for_tests/general.py
@@ -1,0 +1,11 @@
+from sys import stdout
+
+# -----------------------------------------------
+
+def pause():
+
+  stdout.write('Press [return] to continue...')
+  stdout.flush()
+  raw_input()
+
+# -----------------------------------------------
